move asides to own files
